Added animation on poll bars after voting. View poll post finished.

+++ messaging/UIViewController+Flurry.h
@@ -15,6 +15,6 @@
 
 // Can be used to track viewWillAppear
 - (void)sendViewToFlurry:(NSString *)view;
--(void)sendView:(NSString *)view withId:(int)remoteKey;
+- (void)sendView:(NSString *)view withId:(NSInteger)remoteKey;
 
 @end
